worker as strategy pattern
