+++ simulation/run_parallel.py
@@ -1,21 +1,15 @@
 import collections
-import inspect
 import logging
 import operator
-import os
 import random
-import time
 from datetime import datetime
 
 import rx
 from rx import operators as ops
 
-from algorithms.base.driver import BudgetRun
 from evotools import rxtools
-from simulation import factory, log_helper
-from simulation.run_config import NotViableConfiguration
-from simulation.serialization import RunResult
-from simulation.timing import log_time, process_time
+from simulation import factory, worker
+from simulation.timing import log_time
 from simulation.timing import system_time
 
 logger = logging.getLogger(__name__)
@@ -45,7 +39,8 @@ def run_parallel(args):
             log_simulation_stats(start_time, subres[-1], len(simulation_cases))
 
         rx.from_iterable(range(len(simulation_cases))).pipe(
-            ops.flat_map(lambda i: rxtools.from_process(worker, simulation_cases[i], i))
+            ops.map(lambda i: worker.BudgetWorker(simulation_cases[i], i)),
+            ops.flat_map(lambda w: rxtools.from_process(w.run)),
         ).pipe(ops.do_action(on_next=process_result)).run()
     log_summary(args, results, simulation_cases, wall_time)
     rxtools.shutdown_default_executor()
@@ -110,95 +105,3 @@ def log_summary(args, results, simulation_cases, wall_time):
                     **locals()
                 )
             )
-
-
-def worker(simulation, simulation_id):
-    log_helper.init()
-    logger = logging.getLogger(__name__)
-    logger.debug(
-        "Starting the worker. PID: %d, simulation case: %s", os.getpid(), simulation
-    )
-
-    if simulation.renice and os.name == "posix":
-        logger.debug("Renice the process PID:%s by %s", os.getpid(), simulation)
-        os.nice(int(simulation.renice)) # pylint: disable=no-member
-
-    logger.debug("Getting random seed")
-    # basically we duplicate the code of https://github.com/python/cpython/blob/master/Lib/random.py#L111 because
-    # in case os.urandom is not available, random.seed defaults to epoch time. That would set the seed equal in each
-    # process, which is not acceptable.
-    try:
-        random_seed = int.from_bytes(os.urandom(2500), "big")
-    except NotImplementedError:
-        random_seed = int(
-            time.time() * 256 + os.getpid()
-        )  # that's not enough for MT, but will have to do for now.
-    random.seed(random_seed)
-
-    runres = RunResult(simulation)
-
-    try:
-        final_driver, problem_mod = factory.prepare(
-            simulation.algorithm_name, simulation.problem_name
-        )
-
-        logger.debug("Creating the driver used to perform computation")
-
-        driver = final_driver()
-        total_cost, result = 0, None
-
-        proc_time = []
-        results = []
-
-        logger.debug("Beginning processing of %s, simulation: %s", driver, simulation)
-        with log_time(
-            process_time,
-            logger,
-            "Processing done in {time_res}s CPU time",
-            out=proc_time,
-        ):
-
-            def process_results(budget: int):
-                finalpop = driver.finalized_population()
-                finalpop_fit = [
-                    [fit(x) for fit in problem_mod.fitnesses] for x in finalpop
-                ]
-                runres.store(budget, driver.cost, finalpop, finalpop_fit)
-                results.append((driver.cost, finalpop))
-
-            driver.max_budget = simulation.budgets[-1]
-
-            for budget in simulation.budgets:
-                budget_run = BudgetRun(budget)
-                budget_run.create_job(driver).pipe(
-                    ops.do_action(on_completed=lambda: process_results(budget))
-                ).subscribe(
-                    lambda proxy: logger.debug(
-                        "{}{} : Driver progress: budget={}, current cost={}, driver step={}".format(
-                            simulation.algorithm_name,
-                            simulation.problem_name,
-                            budget,
-                            proxy.cost,
-                            proxy.step_no,
-                        )
-                    )
-                )
-
-        return results, proc_time[-1], simulation_id
-
-    except NotViableConfiguration as e:
-        reason = inspect.trace()[-1]
-        logger.info(
-            "Configuartion disabled by %s:%d:%s. simulation case:%s",
-            reason[1],
-            reason[2],
-            reason[3],
-            simulation,
-        )
-        logger.debug("Configuration disabled args:%s. Stack:", exc_info=e)
-
-    except Exception as e:
-        logger.exception("Some error", exc_info=e)
-
-    finally:
-        logger.debug("Finished processing. simulation case:%s", simulation)
